fixed ApiParse module, imports characters and houses
