+++ config/environment.rb
@@ -3,7 +3,6 @@ Bundler.require
 
 ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: 'db/development.db')
 require_all 'lib'
-require_all 'db'
 require 'open-uri'
 require 'json'
